Move read_article wrapper to lib for easier reuse
